chore: moved tests to ./test directory
