+++ jest.config.js
@@ -117,9 +117,7 @@ module.exports = {
   // rootDir: undefined,
 
   // A list of paths to directories that Jest should use to search for files in
-  roots: [
-    'src'
-  ],
+  roots: ['test'],
 
   // Allows you to use a custom runner instead of Jest's default test runner
   // runner: "jest-runner",
