+++ src/helpers/array-helpers.ts
@@ -1,4 +1,4 @@
-export const shuffleArray = (arr: any[]) =>
+export const shuffleArray = (arr: any[]): any[] =>
   arr
     .map(a => [Math.random(), a])
     .sort((a, b) => a[0] - b[0])
