+++ src/types.ts
@@ -14,6 +14,7 @@ declare global {
     }
   }
 }
+
 /**
  * Workaround to make TS recognize this file as a module.
  * https://fettblog.eu/typescript-augmenting-global-lib-dom/
